:recycle: Refactor CMake to use best practices
The changes involved here include:
his
Using GNUInstallDirs for some (but not all) install paths

Using generator expressions supported from 3.4 onwards

Using the APPEND_STRING setting for set_property so that we do not
overwrite the CMAKE_EXE_LINKER_FLAGS variable.

Collated ALL options (dependent or otherwise) into one location

Modify build options to be "namespaced" to prevent collisions when used
via add_subdirectory

Clean up find_package(Python...) so that an actual target is used,
rather than a variable

Moved most file(WRITE) commands to file(GENERATE). This reduces a little
bit of overhead during the configure stage.

Fix having to output informational target into the root binary dir. We
now use file(GENERATE) and $<TARGET_FILE> to get the correct path to
the target.

If git submodules update --init has NOT been run, and the FetchContent
cmake module is available, test and example dependencies will be cloned
as necessary via FetchContent. This will not affect the source
directory, but will place the git clones into _deps by default (This is
configurable with FetchContent. Please see the cmake documentation for
more info)

if `CONFIGURE_DEPENDS` is available for CLI11_headers, it will now be
used. This could help in the future for adding or removing headers from
the current library.

CTest is now provided by default, so that BUILD_TESTING is not defined,
and enable_testing is not manually called. CTest takes care of this for
us.

Removed unnecessary CPack variables, as it will use PROJECT_VERSION by
default.