CI: use the modern Meson `dependency()` method for subprojects
This models the upstream recommendation that consumers avoid the
`subproject('CLI11').get_variable('dep_cli11')` pattern, and instead
simply use `dependency('CLI11')`. This allows build scripts to switch
between system provided dependencies found with pkg-config or cmake, and
subprojects configured and built with the main project without
modification to the build scripts. Users can use the Meson options
`--wrap-mode` and `--force-fallback-for` to get the behavior that they
want.

I have added the `--force-fallback-for=CLI11` switch to the CI to ensure
that the test continues to exercise what it's intended to.