Use pymod.find_installation() to find python in meson.build (#1076)
Since Monterey 12.3, macOS removed python2.7 from it's base
installation. This means that `python` is not longer a command by
default (just `python3`), and the build for CLI11 fails. This PR moves
to using `pymod.find_installation()` to handle this in a more general
way. See https://mesonbuild.com/Python-module.html