formatter: add long_option_alignment_ratio
The default formatter has hardcoded ratio at which the long options are
aligned.  It's currently 1/3 of the column, which makes the default look
awkward:

  -h,     --help              Print
          --option            Something

A ->long_option_alignment_ratio(6/30.f) allows output to look like this:

  -h, --help                  Print
      --option                Something

The 1/3 ratio is also bad if you want to print "descriptive" long
options on a single line, because then you might want to increase the
column width, but that means you waste more space on short options.

e.g. ->column_width(46)

  -l,          --very-descriptive-long-option  Something

vs. ->column_width(38)

  -l,       --very-descriptive-long-option
                                      Something

vs. ->column_width(38) ->long_option_alignment_ratio(6/38.f)

  -l, --very-descriptive-long-option  Something

Any absolute offset `X` can be set as `X/column_width`, so provide a
ratio-based interface.

I would have prefered to give an absolute integer offset, but we still
have to preserve the functionality that does 1/3 if user changed
nothing, which means that ratio-based interface is simpler.

I don't have a good idea for the name, "short_option_ratio" might work
as well.
The setter does not sanity check that the value is in [0;1] range.

Signed-off-by: Radim Krčmář <radim@krcmar.dev>